Adding TERA.twb with double Y plot changes.
There seems to still be some duplicate worksheets created when I imported my workbook. Additionally, the TERA workbook did not load any of the actual data, as it said that the UnionExtract file could not be found. It seems like the location of the UnionExtract file was hardcoded into somewhere on Brennen's machine.